Add GivingTree terrain auto-fix and harden pooled VFX teardown.
Extract terrain prototype cleanup into a dedicated editor menu tool, disable Cartoon FX auto-destruct scripts on pooled one-shot VFX, and centralize Destroy pool preparation. Ignore agent debug session logs.
