+++ .gitignore
@@ -76,3 +76,6 @@ crashlytics-build.properties
 !.ci/*.csproj
 .ci/bin/
 .ci/obj/
+
+# Agent debug session logs
+debug-*.log
