+++ Assets/Editor/GivingTreeTerrainAutoFix.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8f4a2e19b3d4c5a6e7f8091a2b3c4d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
